Moved pseudo-instruction tests to a separate test file.
