+++ src/parser.rs
@@ -4,3 +4,4 @@ pub mod assembling;
 pub mod parser_assembler_main;
 pub mod parser_structs_and_enums;
 pub mod parsing;
+pub mod pseudo_instruction_parsing;
